// src/types/repository.ts
export interface IOwner {
  login: string;
  id: number;
  node_id: string;
  avatar_url: string;
  gravatar_id: string;
  url: string;
  html_url: string;
  followers_url: string;
  following_url: string;
  gists_url: string;
  starred_url: string;
  subscriptions_url: string;
  organizations_url: string;
  repos_url: string;
  events_url: string;
  received_events_url: string;
  type: string;
  user_view_type: string;
  site_admin: boolean;
}

export interface IRepository {
  id: number;
  node_id: string;
  name: string;
  full_name: string;
  private: boolean;
  owner: IOwner;
  html_url: string;
  description: string;
}

export interface IFetchRepositoriesPayload {
  since?: number;
}

export interface ISearchRepositoriesPayload {
  query: string;
}

export interface ILinkHeader {
  next?: string;
  prev?: string;
  first?: string;
  last?: string;
}